Bump to 0.3.1, refactor class to io package
